bankappv2

+++ main.py
@@ -13,17 +13,20 @@ conn = engine.connect()
 def index():
     users = conn.execute(text("SELECT * FROM Accounts ORDER BY AccountID")).all()
     admins = conn.execute(text("SELECT * FROM Accounts")).all()
-    return render_template('index.html', users=users, admins=admins)
+    conn.execute(text("TRUNCATE TABLE Login"))
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    return render_template('index.html', users=users, admins=admins, login=login)
 
 @app.route('/users')
 def get_users():
     users = conn.execute(text(f"SELECT * FROM Accounts ORDER BY AccountID")).all()
+    login = conn.execute(text("SELECT * FROM Login")).first()
     print(users)
-    return render_template('users.html', users=users)
+    return render_template('users.html', users=users, login=login)
 @app.route('/profile', methods=['GET'])
 def get_profile_get_request():
-    
-    return render_template('logging.html', success=False)
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    return render_template('logging.html', success=False, login=login)
 @app.route('/profile', methods=['POST'])
 def get_profile_from_login():
     #logic for login
@@ -31,10 +34,15 @@ def get_profile_from_login():
     password = request.form['password']
     # Check if the username and password are correct
     user = conn.execute(text("SELECT * FROM Accounts WHERE Username = :username"), {'username': username}).first()
+    admin = conn.execute(text("SELECT * FROM Admins WHERE Username = :username"), {'username': username}).first()
+    login = conn.execute(text("SELECT * FROM Login")).first()
     if user:
+        success = False
         # Check if the password matches the hashed password
         if decrypt_password(user.Password_hash, password):
             # Update the Login table to store the login information
+            if user.approved == False:
+                return render_template('logging.html', success=False, username=username, approved = False, admin=False)
             conn.execute(
                 text("TRUNCATE TABLE Login")
             )
@@ -42,27 +50,97 @@ def get_profile_from_login():
                 text("INSERT INTO Login (Username, Password_hash, AdminAccount) VALUES (:username, :password, False)"),
                 {'username': username, 'password': user.Password_hash}
             )
+            login = conn.execute(text("SELECT * FROM Login")).first()
+            print(login)
             # Password is correct, redirect to profile page
-            return render_template('logging.html', success=True, username=username)
+            success = True
+        return render_template('logging.html', success=success, username=username, approved=True, admin=False, login=login)
+    elif admin:
+        if decrypt_password(admin.Password_hash, password):
+            # Update the Login table to store the login information
+            conn.execute(
+                text("TRUNCATE TABLE Login")
+            )
+            conn.execute(
+                text("INSERT INTO Login (Username, Password_hash, AdminAccount) VALUES (:username, :password, True)"),
+                {'username': username, 'password': admin.Password_hash}
+            )
+            login = conn.execute(text("SELECT * FROM Login")).first()
+            # Password is correct, redirect to profile page
+            return render_template('logging.html', success=True, username=username, approved=True, admin=True, login=login)
         else:
-            # Password is incorrect, show error message
-            return render_template('logging.html', success=False, username=username)
-    return render_template('logging.html', success=False)
+            return render_template('logging.html', success=False, username=username, approved=True, admin=False, login=login)
 @app.route('/profile/<username>', methods=['GET'])
 def get_profile(username=None):
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    admins = conn.execute(text("SELECT * FROM Admins")).all()
     if id != None:
         users = conn.execute(text("SELECT * FROM Accounts WHERE Username = :username"), {'username': username}).first()
-        return render_template('profile.html', users=users, failure=False)
+        return render_template('profile.html', users=users, failure=False, login=login)
     else: 
         users = conn.execute(text(f"SELECT * FROM Accounts ORDER BY AccountID")).all()
         print(users)
-        return render_template('index.html', users=users, failure=True)
+        return render_template('index.html', users=users, failure=True, login=login, admins=admins)
 
-@app.route('/create', methods=['GET'])
-def create_get_request():
-    return render_template('boats_create.html')
-
-
+@app.route('/admin', methods=['GET'])
+def admin_get_request():
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    users = conn.execute(text(f"SELECT * FROM Accounts ORDER BY AccountID")).all()
+    admins = conn.execute(text("SELECT * FROM Admins")).all()
+    return render_template('admin.html', users=users, admins=admins, login=login, success = None)
+@app.route('/admin', methods=['POST'])
+def admin_post_request():
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    admins = conn.execute(text("SELECT * FROM Admins")).all()
+    admin = conn.execute(text("SELECT * FROM Admins WHERE Username = :username"), {'username': login.Username}).first()
+    if request.form['approved'] == '1':
+        # Approve user
+        # Save previus value of approved
+        previous_approved = conn.execute(
+            text("SELECT approved FROM Accounts WHERE AccountID = :id"),
+            request.form
+        ).first()
+        conn.execute(
+            text("UPDATE Accounts SET approved = True WHERE AccountID = :id"),
+            request.form
+        )
+        new_approved = conn.execute(
+            text("SELECT approved FROM Accounts WHERE AccountID = :id"),
+            request.form
+        ).first()
+        #update the admin transcript table
+        conn.execute(
+            text(f"INSERT INTO AdminTranscript (AdminUsername, actionDate, accountID, rowChange, previousValue, newValue) VALUES ('{str(admin.Username)}', NOW(), :account_id, 'approved','{str(previous_approved)}', '{str(new_approved)}')"),
+            {'account_id': request.form['id']}
+        )
+    elif request.form['approved'] == '0':
+        # Reject user
+        # Save previus value of approved
+        previous_approved = conn.execute(
+            text("SELECT approved FROM Accounts WHERE AccountID = :id"),
+            request.form
+        ).first()
+        conn.execute(
+            text("UPDATE Accounts SET approved = False WHERE AccountID = :id"),
+            request.form
+        )
+        new_approved = conn.execute(
+            text("SELECT approved FROM Accounts WHERE AccountID = :id"),
+            request.form
+        ).first()
+        #update the admin transcript table
+        conn.execute(
+            text(f"INSERT INTO AdminTranscript (AdminUsername, actionDate, accountID, rowChange, previousValue, newValue) VALUES ('{str(admin.Username)}', NOW(), :account_id, 'approved','{str(previous_approved)}', '{str(new_approved)}')"),
+            {'account_id': request.form['id']}
+        ) 
+    users = conn.execute(text(f"SELECT * FROM Accounts ORDER BY AccountID")).all()
+    return render_template('admin.html', users=users, admins=admins, login=login, success="Updated successfully!")
+@app.route('/admin/transcript', methods=['GET'])
+def transcript_get_request():
+    login = conn.execute(text("SELECT * FROM Login")).first()
+    admins = conn.execute(text("SELECT * FROM Admins")).all()
+    transcripts = conn.execute(text("SELECT * FROM AdminTranscript")).all()
+    return render_template('admin_transcript.html', transcript=transcripts, login=login, admins=admins)
 @app.route('/create', methods=['POST'])
 def create_boat():
     # you can access the values with request.from.name
